backend/requirements.txt
# Works on Python 3.10–3.14 without Rust (uses Pydantic v1).
# For Pydantic v2 + latest FastAPI, use Python 3.11 or 3.12.
fastapi>=0.100.0,<0.105.0
uvicorn[standard]>=0.27.0
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.9
pydantic>=1.10.0,<2.0.0
email-validator>=2.0.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.6
# Optional: for "fastapi dev main.py" (else use: uvicorn main:app --reload)
fastapi-cli>=0.0.8
